requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
numpy==2.2.6
packaging==25.0
pandas==2.3.1
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Unidecode==1.4.0
whitenoise==6.9.0
